task7/requirements.txt
deeppavlov
python-telegram-bot
python-weather
fasttext
tensorflow
